+++ eosets/__init__.py
@@ -17,18 +17,9 @@
 """
 **EOSets** library
 """
+
 # flake8: noqa
-from .__meta__ import (
-    __author__,
-    __author_email__,
-    __copyright__,
-    __description__,
-    __documentation__,
-    __license__,
-    __title__,
-    __url__,
-    __version__,
-)
+from .__meta__ import __version__
 
 EOSETS_NAME = "eosets"